+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.5] - 2022-09-03
+### Added
+- Support for forced installation
+### Fixed
+- Formulae showing up as casks
+
 ## [0.0.4] - 2022-07-02
 ### Added
 - Additional tap metadata retrieval